api/requirements.txt
fastapi==0.62.0
pydantic==1.7.3
SQLAlchemy==1.3.22
starlette==0.13.6
uvicorn
psycopg2
itsdangerous
passlib
python-multipart
python-jose